:hammer: Move admin pages under /admin route
