+++ frontend/src/pages/Home.tsx
@@ -4,7 +4,6 @@
 
 import {Container} from "@mui/material"
 import * as React from "react"
-import Page from "./Page"
 import {UndrawUnderConstruction} from "../components/illustrations"
 
 /**
@@ -13,11 +12,11 @@ import {UndrawUnderConstruction} from "../components/illustrations"
  */
 export default function HomePage(): JSX.Element {
   return (
-    <Page title="Home">
-      <Container className="flex justify-center items-center h-full flex-col mt-1">
-        <UndrawUnderConstruction className="w-1/2 h-1/2" />
-        <h3>This page is under construction</h3>
-      </Container>
-    </Page>
+    <Container
+      className="p-0 h-screen flex justify-center items-center flex-col"
+      maxWidth={false}>
+      <UndrawUnderConstruction className="w-1/2 h-1/2" />
+      <h3>This page is under construction</h3>
+    </Container>
   )
 }
